nfc: set right the nfcti_enhancement
FCTI_ENHANCEMENT property setting should flow in from the system.prop file
and should not be by default defined as true.
If it is defined to be true, then there is no point in introducing the change
itself, since it is set to true to begin with.

The flag in itself is required to avoid certain sections of code in Nfc.apk
that are specific to nxp chipset originally from Google.
So, in packages/apps/Nfc/ we should check for
!SystemProperties.NFCTI_ENHANCEMENT and avoid nxp specific sections of code.

Change-Id: I6a4bdc257b4ec39429438c793df30dd2671fdd3d
Signed-off-by: bvijay <bvijay@ti.com>